+++ srcs/I-O/ft_ps.c
@@ -6,30 +6,30 @@
 /*   By: adelille <adelille@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/26 13:09:31 by adelille          #+#    #+#             */
-/*   Updated: 2021/11/22 20:33:38 by adelille         ###   ########.fr       */
+/*   Updated: 2022/01/06 18:32:42 by adelille         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/libft.h"
 
-int	ft_ps(char *str)
+ssize_t	ft_ps(char *str)
 {
 	return (write(STDOUT, str, ft_strlen(str)));
 }
 
-int	ft_psc(char *str, char *color)
+ssize_t	ft_psc(char *str, char *color)
 {
 	return (write(STDOUT, color, ft_strlen(color))
 		+ write(STDOUT, str, ft_strlen(str))
 		+ write(STDOUT, DEF, ft_strlen(DEF)));
 }
 
-int	ft_pser(char *str)
+ssize_t	ft_pser(char *str)
 {
 	return (write(STDERR, str, ft_strlen(str)));
 }
 
-int	ft_pserc(char *str, char *color)
+ssize_t	ft_pserc(char *str, char *color)
 {
 	return (write(STDERR, color, ft_strlen(color))
 		+ write(STDERR, str, ft_strlen(str))
